backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.13",
    "langchain-groq>=1.1.2",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn>=0.42.0",
    "elevenlabs>=2.3.0",
]
